% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2.R
\name{convert_hmdb2fmsea}
\alias{convert_hmdb2fmsea}
\title{Convert metpath HMDB object to fmsea database format}
\usage{
convert_hmdb2fmsea(hmdb_pathway)
}
\arguments{
\item{hmdb_pathway}{An HMDB pathway object from the metpath package}
}
\value{
A data frame formatted for featuremsea/fmsea
}
\description{
Convert metpath HMDB object to fmsea database format
}
