% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2.R
\name{convert_kegg2fmsea}
\alias{convert_kegg2fmsea}
\title{Convert metpath KEGG object to fmsea database format}
\usage{
convert_kegg2fmsea(kegg_hsa_pathway)
}
\arguments{
\item{kegg_hsa_pathway}{A KEGG pathway object from the metpath package}
}
\value{
A data frame formatted for featuremsea/fmsea
}
\description{
Convert metpath KEGG object to fmsea database format
}
