% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2.R
\name{convert_wikipathway2fmsea}
\alias{convert_wikipathway2fmsea}
\title{Convert WikiPathways Object to featuremsea Format}
\usage{
convert_wikipathway2fmsea(wiki_pathway)
}
\arguments{
\item{wiki_pathway}{An S4 object containing WikiPathways data. It is expected to have
slots for `describtion`, `pathway_class`, `compound_list` (containing HMDB.ID and KEGG.ID),
`pathway_id`, and `pathway_name`.}
}
\value{
A data frame containing the following columns:
\itemize{
  \item pathway_id
  \item pathway_name
  \item pathway_description
  \item HMDB_ID (Collapsed unique IDs)
  \item KEGG_ID (Collapsed unique IDs)
  \item database (Set to "WikiPathways")
  \item pathway_class_all
}
}
\description{
This function extracts pathway information from a WikiPathways S4 object and formats it
into a data frame suitable for the featuremsea package.
It assumes the compound list has already been mapped to HMDB and KEGG IDs.
}
