% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_featuremsea_object.R
\docType{class}
\name{featuremsea_object-class}
\alias{featuremsea_object-class}
\title{featuremsea_object Class}
\description{
An S4 class to store the results of feature-based MSEA analysis.
}
\section{Slots}{

\describe{
\item{\code{feature_metabolite_count}}{A data frame of feature-metabolite counts.}

\item{\code{annotation_table_weighting}}{A data frame of weighted annotations.}

\item{\code{significant_modules}}{A data frame of significant modules.}

\item{\code{res_list}}{A list containing raw results.}

\item{\code{converged}}{Logical. Whether the algorithm converged.}

\item{\code{iterations_used}}{Integer. Number of iterations used.}

\item{\code{process_info}}{A list containing processing timestamp/info.}
}}

