% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_featuremsea_object.R
\name{show,featuremsea_object-method}
\alias{show,featuremsea_object-method}
\title{Show method for featuremsea_object}
\usage{
\S4method{show}{featuremsea_object}(object)
}
\arguments{
\item{object}{An object of class featuremsea_object.}
}
\description{
Show method for featuremsea_object
}
